led_controller/raspberry_pi/pyproject.toml
[project]
name = "whale-sounds"
version = "0.1.0"
description = "Simple ambient synth for the Gourd"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ajsonrpc==1.2.0",
    "anyio==4.9.0",
    "bottle==0.13.4",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "gpiozero>=2.0.1", # TODO: remove if not reading GPIO directly into RasPi
    "h11==0.16.0",
    "idna==3.10",
    "keyboard>=0.13.5",
    "lgpio>=0.2.2.0", # TODO: remove if not reading GPIO directly into RasPi
    "marshmallow==3.26.1",
    "packaging==25.0",
    "pigpio>=1.78",
    "platformio==6.1.18",
    "pyelftools==0.32",
    "pyo",
    "pyserial==3.5",
    "requests==2.32.4",
    "rpi-gpio>=0.7.1", # TODO: remove if not reading GPIO directly into RasPi
    "semantic-version==2.10.0",
    "sniffio==1.3.1",
    "spidev>=3.7", # TODO: remove if not reading GPIO directly into RasPi
    "starlette==0.46.2",
    "tabulate==0.9.0",
    "typing-extensions==4.14.1",
    "urllib3==2.5.0",
    "uvicorn==0.34.3",
    "wsproto==1.2.0",
]

[tool.uv.sources]
pyo = { path = "pyo-1.0.6-cp311-cp311-linux_aarch64.whl" }
